requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# Database
PyMySQL==1.1.0
cryptography==41.0.7

# Environment Management
python-dotenv==1.0.0
python-decouple==3.8

# Scheduling
APScheduler==3.10.4

# Data Processing & Visualization
matplotlib==3.8.2
pandas==2.1.4

# Date Parsing
python-dateutil==2.8.2

# Development Tools
pytest==7.4.3
black==23.12.1
flake8==7.0.0

# Utilities
pytz==2023.3

# Mini App API (FastAPI)
fastapi==0.109.2
uvicorn[standard]==0.27.1
